Labfiles/07-translation/Python/translators/requirements.txt
python-dotenv
azure-identity
azure-ai-translation-text==1.0.1
azure-cognitiveservices-speech==1.48.2